+++ relm4-components/src/lib.rs
@@ -1,11 +1,14 @@
 //! Collection of reusable and easily configurable components for Relm4.
 
-#![warn(missing_docs, rust_2018_idioms, unreachable_pub)]
-#![doc(
-    html_logo_url = "https://raw.githubusercontent.com/AaronErhardt/relm4/main/assets/Relm_logo.svg"
-)]
-#![doc(
-    html_favicon_url = "https://raw.githubusercontent.com/AaronErhardt/relm4/main/assets/Relm_logo.svg"
+#![doc(html_logo_url = "https://relm4.org/icons/relm4_logo.svg")]
+#![doc(html_favicon_url = "https://relm4.org/icons/relm4_org.svg")]
+#![warn(
+    missing_debug_implementations,
+    missing_docs,
+    rust_2018_idioms,
+    unreachable_pub,
+    clippy::cargo,
+    clippy::must_use_candidate
 )]
 
 pub mod alert;