+++ relm4-components/src/simple_combo_box.rs
@@ -107,6 +107,7 @@ where
     }
 
     /// Return the value of the currently selected element or [`None`] if nothing is selected.
+    #[must_use]
     pub fn get_active_elem(&self) -> Option<&E> {
         self.active_index.map(|idx| &self.variants[idx])
     }