+++ relm4-macros/src/widgets/gen/util/if_branch.rs
@@ -4,7 +4,7 @@ use quote::quote;
 use crate::widgets::{IfBranch, IfCondition};
 
 impl IfBranch {
-    pub fn update_stream(
+    pub(crate) fn update_stream(
         &self,
         stream: &mut TokenStream2,
         inner_update_tokens: TokenStream2,